Revert "float CodeMirror expressions so they don't take up the whole horizontal space"
This reverts commit d8a77a8eb9fc43d432104f4698e466764f90b211.

I just updated to Chrome 53.0.2785.116 and now Apparatus seizes up and
hits 100% CPU before anything is rendered. I used `git bisect` to
track down the problem, and mind-bogglingly, it was this one. Testing
in DevTools shows that the problem is setting .CodeMirror to "float:
left". It used to work perfectly, now it's broken. Fact is, I have no
idea how to debug this problem, and this change isn't essential. So,
farewell, d8a77a8e.