Generic flexbox layout classes (+ assorted cleanup)
My hypothesis is that, as far as app UI layout is concerned, the
HTML/CSS "separation of concerns" is absurd. Rather than scatter
`flex` properties throughout our CSS to specify layout behavior, let's
design a set of reusable layout primitives which we can refer to
directly in our HTML. So far, it looks nice.